+++ src/commands/index.ts
@@ -117,6 +117,7 @@ import { zulu } from './utils/zulu';
 import { latlongfix } from './general/latlongfix';
 import { headwind } from './general/headwind';
 import { wolframalpha } from './utils/wolframalpha';
+import { birthday } from './utils/birthday';
 import { recommendedSettings } from './a32nx/recommendedsettings';
 
 const commands: CommandDefinition[] = [
@@ -237,6 +238,7 @@ const commands: CommandDefinition[] = [
     latlongfix,
     headwind,
     wolframalpha,
+    birthday,
     recommendedSettings,
 ];