+++ src/index.ts
@@ -7,6 +7,7 @@ import { join } from 'path';
 import commands from './commands';
 import { makeEmbed } from './lib/embed';
 import Logger from './lib/logger';
+import { connect } from './lib/db';
 
 dotenv.config();
 const apm = start({
@@ -27,6 +28,12 @@ let healthy = false;
 client.on('ready', () => {
     Logger.info(`Logged in as ${client.user.tag}!`);
     healthy = true;
+
+    // Connect to database
+    if (process.env.MONGODB_URL) {
+        connect(process.env.MONGODB_URL)
+            .catch(Logger.error);
+    }
 });
 
 client.on('disconnect', () => {