+++ .github/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 Update <small>_ May 2022</small>
 
+- feat: add birthday command (12/05/2022)
 - refactor: weight and balance command (11/05/2022)
 - feat: recommended settings command (11/05/2022)
 - docs(exp): Update support wording (09/05/2022)